requirements.txt
python-dotenv==1.0.0
openai==1.12.0
scikit-learn==1.4.0
together==0.2.5
seaborn==0.13.2
matplotlib==3.8.2
numpy==1.26.4
pandas==2.2.0
pathlib==1.0.1
